Ajute migration

+++ database/migrations/2020_07_25_115313_alter_collun_name_sfdoccontabilizacao_table.php
@@ -13,10 +13,12 @@ class AlterCollunNameSfdoccontabilizacaoTable extends Migration
      */
     public function up()
     {
-        Schema::table('sfdoccontabilizacao', function (Blueprint $table) {
-            $table->dropColumn('numcodcont');
-            $table->string('numdoccont')->nullable();
-        });
+        if(Schema::hasColumn('sfdoccontabilizacao','numcodcont')){
+            Schema::table('sfdoccontabilizacao', function (Blueprint $table) {
+                $table->dropColumn('numcodcont');
+                $table->string('numdoccont')->nullable();
+            });
+        }
     }
 
     /**
